app router + polish + new photo
