+++ package.json
@@ -15,18 +15,19 @@
     "@types/node": "18.11.9",
     "@types/react": "18.0.25",
     "@types/react-dom": "18.0.8",
-    "@vercel/analytics": "^0.1.3",
+    "@vercel/analytics": "^1.0.0",
     "eslint": "8.26.0",
     "eslint-config-next": "^15.0.3",
     "next": "^15.0.3",
-    "next-themes": "^0.2.1",
+    "next-themes": "^0.4.3",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
     "typescript": "4.8.4"
   },
   "devDependencies": {
-    "autoprefixer": "^10.4.13",
-    "postcss": "^8.4.18",
-    "tailwindcss": "^3.2.2"
+    "autoprefixer": "^10.4.20",
+    "postcss": "^8.4.49",
+    "prettier": "^2.8.4",
+    "tailwindcss": "^3.4.14"
   }
 }
