+++ src/components/Game/Footer.tsx
@@ -1,4 +1,4 @@
-"use-client";
+"use client";
 
 import Badge from "./Badge";
 import { useOthers } from "@liveblocks/react/suspense";
