+++ src/styles/globals.css
@@ -7,5 +7,5 @@ body {
 }
 
 ::selection {
-  @apply bg-orange-100 text-orange-900 dark:bg-orange-900 dark:text-orange-100;
+  @apply bg-rose-100 text-rose-900 dark:bg-rose-900 dark:text-rose-100;
 }
